+++ packages/button/src/Button/Button.spec.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
-import { fireEvent, render } from '@testing-library/react';
+import { fireEvent, render, screen } from '@testing-library/react';
 import { axe } from 'jest-axe';
 import NextLink from 'next/link';
 
 import { BoxProps } from '@leafygreen-ui/box';
+import { Spinner } from '@leafygreen-ui/loading-indicator';
 
 import { ButtonProps } from '../types';
 import Button from '..';
@@ -52,6 +53,33 @@ describe('packages/button', () => {
         children: child,
       });
       expect(button.textContent).toBe(child);
+    });
+
+    test(`renders spinner when isLoading is true`, () => {
+      const { getByTestId } = renderButton({
+        isLoading: true,
+        loadingIndicator: <Spinner />,
+      });
+      expect(getByTestId('lg-button-spinner')).toBeVisible();
+    });
+
+    test(`does not loadingText when isLoading is false`, () => {
+      const loadingText = 'loading text';
+      const { queryByText } = renderButton({
+        isLoading: false,
+        loadingText,
+      });
+      expect(queryByText(loadingText)).toBeNull();
+    });
+
+    test(`renders loadingText when isLoading is true`, () => {
+      const loadingText = 'loading text';
+      const { getByText } = renderButton({
+        isLoading: true,
+        loadingIndicator: <Spinner />,
+        loadingText,
+      });
+      expect(getByText(loadingText)).toBeVisible();
     });
 
     test(`renders "${title}" as the button title`, () => {
@@ -109,10 +137,10 @@ describe('packages/button', () => {
 
     test(`renders a button as another HTML element if the "as" prop is set`, () => {
       const { container, button } = renderButton({
-        as: 'p',
+        as: 'div',
       });
       expect(container.querySelector('button')).not.toBeInTheDocument();
-      expect(button.tagName.toLowerCase()).toBe('p');
+      expect(button.tagName.toLowerCase()).toBe('div');
     });
 
     test(`renders a when passing in a NextJS Link wrapper`, () => {
@@ -170,6 +198,16 @@ describe('packages/button', () => {
       expect(onClick).toHaveBeenCalledTimes(0);
     });
 
+    test('does not fire onClick handler when loading', () => {
+      const onClick = jest.fn();
+      const { button } = renderButton({
+        isLoading: true,
+        onClick,
+      });
+      fireEvent.click(button);
+      expect(onClick).toHaveBeenCalledTimes(0);
+    });
+
     test('does not fire onClick handler on disabled anchor', () => {
       const onClick = jest.fn();
       const { button } = renderButton({
@@ -200,6 +238,22 @@ describe('packages/button', () => {
       });
       expect(button).toHaveAttribute('href', href);
     });
+
+    test('does not invoke a forms submit handler when disabled', () => {
+      const onSubmit = jest.fn();
+
+      render(
+        <form onSubmit={onSubmit}>
+          <Button disabled type="submit">
+            Submit
+          </Button>
+        </form>,
+      );
+
+      const button = screen.getByRole('button');
+      fireEvent.click(button);
+      expect(onSubmit).not.toHaveBeenCalled();
+    });
   });
 
   /* eslint-disable jest/no-disabled-tests, jest/expect-expect*/
@@ -213,7 +267,7 @@ describe('packages/button', () => {
     });
 
     test('accepts a string as `as`', () => {
-      <Button as="p" />;
+      <Button as="div" />;
     });
 
     test('accepts a component as `as`', () => {