+++ packages/menu/src/types.ts
@@ -4,3 +4,7 @@ export const Size = {
 } as const;
 
 export type Size = typeof Size[keyof typeof Size];
+
+export type ElementOf<
+  T extends React.ComponentType<React.PropsWithChildren<unknown>>,
+> = React.ReactComponentElement<T, React.ComponentPropsWithRef<T>>;