+++ packages/modal/src/Modal/ModalView.tsx
@@ -24,8 +24,13 @@ import {
   scrollContainer,
   visibleBackdrop,
   visibleModalContentStyle,
-} from './styles';
-import { CloseIconColor, ForwardedRef, ModalProps, ModalSize } from './types';
+} from './Modal.styles';
+import {
+  CloseIconColor,
+  ForwardedRef,
+  ModalProps,
+  ModalSize,
+} from './Modal.types';
 
 /**
  * @internal
@@ -70,15 +75,6 @@ const ModalView = React.forwardRef(
 
     useEscapeKey(handleClose, { enabled: open && !isPopoverOpen });
 
-    const focusTrapOptions = initialFocus
-      ? {
-          initialFocus: `#${id} ${initialFocus}`,
-          fallbackFocus: `#${closeId}`,
-        }
-      : {
-          fallbackFocus: `#${closeId}`, // tests fail without a fallback. (https://github.com/focus-trap/focus-trap-react/issues/91)
-        };
-
     return (
       <Transition
         in={open}
@@ -103,7 +99,12 @@ const ModalView = React.forwardRef(
               )}
             >
               <LeafyGreenProvider darkMode={darkMode}>
-                <FocusTrap focusTrapOptions={focusTrapOptions}>
+                <FocusTrap
+                  active={state === 'entered'}
+                  focusTrapOptions={{
+                    fallbackFocus: `#${closeButton}`,
+                  }}
+                >
                   <div
                     className={scrollContainer}
                     ref={el => setScrollContainerRef(el)}