+++ packages/search-input/src/SearchInput/SearchInput.styles.ts
@@ -4,6 +4,7 @@ import { palette } from '@leafygreen-ui/palette';
 import {
   focusRing,
   fontFamilies,
+  fontWeights,
   hoverRing,
   transitionDuration,
   typeScales,
@@ -127,13 +128,13 @@ export const baseInputStyle = css`
   font-family: ${fontFamilies.default};
   width: 100%;
   height: 1.5em;
-  font-weight: normal;
+  font-weight: ${fontWeights.regular};
   z-index: 1;
   outline: none;
   border: none;
   padding: 0;
 
-  &:disabled {
+  &[aria-disabled='true'] {
     cursor: not-allowed;
 
     &:hover,
@@ -168,7 +169,7 @@ export const inputThemeStyle: Record<Theme, string> = {
       border: none;
       -webkit-text-fill-color: inherit;
 
-      &:not(:disabled) {
+      &[aria-disabled='false'] {
         box-shadow: ${autofillShadowOverride(palette.white)};
 
         &:focus {
@@ -185,10 +186,10 @@ export const inputThemeStyle: Record<Theme, string> = {
 
     &::placeholder {
       color: ${palette.gray.base};
-      font-weight: normal;
+      font-weight: ${fontWeights.regular};
     }
 
-    &:disabled {
+    &[aria-disabled='true'] {
       &::placeholder {
         color: ${palette.gray.base};
       }
@@ -211,7 +212,7 @@ export const inputThemeStyle: Record<Theme, string> = {
       background: transparent;
       border: none;
       -webkit-text-fill-color: ${palette.gray.light3};
-      &:not(:disabled) {
+      &[aria-disabled='false'] {
         box-shadow: ${autofillShadowOverride(palette.gray.dark4)};
 
         &:focus {
@@ -229,12 +230,12 @@ export const inputThemeStyle: Record<Theme, string> = {
 
     &::placeholder {
       color: ${palette.gray.base};
-      font-weight: normal;
+      font-weight: ${fontWeights.regular};
     }
 
-    &:disabled {
+    &[aria-disabled='true'] {
       &::placeholder {
-        color: ${palette.gray.base};
+        color: ${palette.gray.dark1};
       }
 
       &:-webkit-autofill {