+++ packages/emotion/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @leafygreen-ui/emotion
 
+## 4.0.5
+
+### Patch Changes
+
+- 215268ff: Updates build tooling. No functional changes
+
+## 4.0.4
+
+### Patch Changes
+
+- 8c0c2bdf9: Updates build script to include a transpiled copy of the story file in the bundle
+
 ## 4.0.3
 
 ### Patch Changes